\name{faults.math.x.100}
\alias{faults.math.x.100}
\docType{data}
\title{
faults.math.x.100
}
\description{
faults.math.x.100
}
\usage{data("faults.math.x.100")}
\format{
  A data frame with 342 observations on the following 7 variables.
  \describe{
    \item{\code{project}}{a factor with levels \code{math}}
    \item{\code{power}}{a numeric vector}
    \item{\code{number}}{a numeric vector}
    \item{\code{SeparateFaults}}{a numeric vector}
    \item{\code{JoinedFaults}}{a numeric vector}
    \item{\code{RemovedFaults}}{a numeric vector}
    \item{\code{addedfails}}{a numeric vector}
  }
}

\keyword{datasets}
