\name{faults.x.2.10}
\alias{faults.x.2.10}
\docType{data}
\title{
faults.x.2.10
}
\description{
faults.x.2.10
}
\usage{data("faults.x.2.10")}
\format{
  A data frame with 192 observations on the following 7 variables.
  \describe{
    \item{\code{project}}{a factor with levels \code{bcel} \code{cli} \code{codec} \code{collections} \code{compress} \code{configuration} \code{csv} \code{dbcp} \code{exec} \code{fileupload} \code{imaging} \code{io} \code{jexl} \code{lang} \code{mail} \code{math} \code{net} \code{pool} \code{scxml} \code{validator}}
    \item{\code{power}}{a numeric vector}
    \item{\code{number}}{a numeric vector}
    \item{\code{SeparateFaults}}{a numeric vector}
    \item{\code{JoinedFaults}}{a numeric vector}
    \item{\code{RemovedFaults}}{a numeric vector}
    \item{\code{addedfails}}{a numeric vector}
  }
}

\keyword{datasets}
