\name{MutCoupling-package}
\alias{MutCoupling-package}
\alias{MutCoupling}
\docType{package}
\title{
\packageTitle{MutCoupling}
}
\description{
\packageDescription{MutCoupling}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{MutCoupling}
\packageIndices{MutCoupling}
}
\author{
\packageAuthor{MutCoupling}

Maintainer: \packageMaintainer{MutCoupling}
}
\references{
}
\keyword{ package }
\seealso{
}
\examples{
}
